refactor(UserQualifications): updating middleware imports, renaming NAvigation.actions to Navigation.utils
