+++ src/modules/Error/index.ts
@@ -1,4 +1,5 @@
 import * as middleware from './Error.middleware'
 import * as actions from './Error.reducer'
+import * as utils from './error.utils'
 
-export { actions, middleware }
+export { actions, middleware, utils }
