+++ src/modules/Navigation/index.ts
@@ -1,3 +1,5 @@
 import Navigation from './Navigation.container'
+import * as utils from './Navigation.utils'
 
 export default Navigation
+export { utils }
