+++ src/modules/UserQualifications/UserQualifications.middleware.ts
@@ -1,41 +1,36 @@
 import { mergeRight } from 'ramda'
 import { Middleware } from 'redux'
 
-import { actions as ApiActions, utils as ApiUtils } from '../../api'
-import { constants as ApiUsersConstants } from '../../api/users'
-import { UserCredentialTypes } from '../../api/users/users.types'
-import { NormaliseDependency } from '../../redux/redux.types'
-import { extractDataFromResponseAction } from '../../redux/redux.utils'
-import { StdFn } from '../../types/general.types'
-import { showSimpleMessage } from '../../utils/error'
-import { extractErrorResponseMessage } from '../Error/error.utils'
+import { actions as ApiActions, utils as ApiUtils } from '~/api'
+import { constants as ApiUsersConstants, types as ApiUsersTypes } from '~/api/users'
+import { NormaliseDependency } from '~/redux/redux.types'
+import * as ReduxUtils from '~/redux/redux.utils'
+import * as GeneralTypes from '~/types/general.types'
+import { showSimpleMessage } from '~/utils/error'
+
+import { utils as ErrorUtils } from '../Error'
 import { HomeNavigationRoutes } from '../HomeNavigation/HomeNavigation.types'
-import * as Navigation from '../Navigation/Navigation.actions'
-import * as QualificationsActions from '../Qualifications/Qualifications.reducer'
-import * as UserActions from '../User/User.reducer'
-import * as UserSelectors from '../User/User.selector'
-import { UserCredentials } from '../User/User.types'
-import { extractUserCredentialFormValues, prepareUserCredentialItemPayload } from '../User/User.utils'
+import { utils as NavigationUtils } from '../Navigation'
+import { actions as QualificationsActions } from '../Qualifications'
+import { actions as UserActions, selectors as UserSelectors, types as UserTypes, utils as UserUtils } from '../User'
 import {
   clearUserQualificationFormValues,
   createUserQualification,
   createUserQualificationFailure,
   createUserQualificationSuccess,
-  setUserQualificationFormValues,
-  updateUserQualifications,
-} from './UserQualifications.reducer'
-import {
   getUserQualificationsSuccess,
   normaliseUserQualificationsSuccess,
+  setUserQualificationFormValues,
   setUserQualifications,
+  updateUserQualifications,
 } from './UserQualifications.reducer'
 import { selectFormValues } from './UserQualifications.selector'
 import { UserQualification } from './UserQualifications.types'
 
 export const getUserQualificationsFromCredentialsFlow =
   (
-    extractDataFromPayload: StdFn<any, UserCredentials>,
-    extractQualifications: StdFn<UserCredentials, UserQualification[]>,
+    extractDataFromPayload: GeneralTypes.StdFn<any, UserTypes.UserCredentials>,
+    extractQualifications: GeneralTypes.StdFn<UserTypes.UserCredentials, UserQualification[]>,
   ): Middleware =>
   ({ dispatch }) =>
   next =>
@@ -79,7 +74,9 @@ export const setUserQualificationsFormValuesFlow: Middleware =
   action => {
     const result = next(action)
     if (QualificationsActions.createQualification.match(action)) {
-      const formValues = extractUserCredentialFormValues(UserCredentialTypes.Qualification)(action.payload)
+      const formValues = UserUtils.extractUserCredentialFormValues(ApiUsersTypes.UserCredentialTypes.Qualification)(
+        action.payload,
+      )
       dispatch(setUserQualificationFormValues(formValues))
     }
     return result
@@ -96,7 +93,7 @@ export const createUserQualificationFlow: Middleware =
       const userId = UserSelectors.selectId(state)
 
       const formValues = selectFormValues(state)
-      const userQualificationsPayload = prepareUserCredentialItemPayload(action)(formValues)
+      const userQualificationsPayload = UserUtils.prepareUserCredentialItemPayload(action)(formValues)
 
       const config = ApiUtils.prependValueToEndpointInConfig(ApiUsersConstants.USERS_CREDENTIALS_CREATE_CONFIG)(userId)
 
@@ -120,11 +117,11 @@ export const createUserQualificationSuccessFlow =
   action => {
     const result = next(action)
     if (createUserQualificationSuccess.match(action)) {
-      const data = extractDataFromResponseAction(action)
+      const data = ReduxUtils.extractDataFromResponseAction(action)
       const normalised = normalise([data])
       dispatch(updateUserQualifications(normalised))
       dispatch(clearUserQualificationFormValues())
-      Navigation.navigate(HomeNavigationRoutes.Home)
+      NavigationUtils.navigate(HomeNavigationRoutes.Home)
     }
     return result
   }
@@ -137,7 +134,7 @@ export const createUserQualificationFailureFlow =
     const result = next(action)
 
     if (createUserQualificationFailure.match(action)) {
-      const errorMessage = extractErrorResponseMessage(action)
+      const errorMessage = ErrorUtils.extractErrorResponseMessage(action)
       // TODO: this should be handled by the notification module
       notification('danger', 'Error', errorMessage)
     }
